Use `null` write fn when operator codegen errors
Since `HydroflowGraph::as_code()` always returns the codegen regardless
of errors, we should do our best to make the returned code as sensible
as possible in the face of errors.
